frequirements.txt
accelerate==0.30.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
asgiref==3.9.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
billiard==4.2.1
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
celery==5.3.6
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chroma-hnswlib==0.7.3
chromadb==0.4.24
clarifai==10.11.1
clarifai-grpc==11.7.9
clarifai-protocol==0.0.14
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cohere==5.17.0
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
crewai==0.30.11
crewai-tools==0.1.7
cryptography==45.0.6
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.15
durationpy==0.10
embedchain==0.1.113
exceptiongroup==1.3.0
fastapi==0.110.3
fastavro==1.12.0
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.111.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.31.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
inquirerpy==0.3.4
instructor==0.5.2
intel-openmp==2021.4.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
kubernetes==33.1.0
lancedb==0.5.7
langchain==0.1.20
langchain-cohere==0.1.5
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.147
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mkl==2021.4.0
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.6.4
mutagen==1.47.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.102.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.11.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.4
pfzy==0.3.4
pillow==10.3.0
pluggy==1.6.0
posthog==3.25.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pulsar-client==3.8.0
py==1.11.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==1.59.1
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.9.18
PyNaCl==1.5.0
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.404
pysbd==0.3.4
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-rapidjson==1.21
pytube==15.0.0
PyYAML==6.0.2
ratelimiter==1.2.0.post0
redis==5.0.8
regex==2023.12.25
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.4
rsa==4.9.1
safetensors==0.6.2
schema==0.7.5
selenium==4.35.0
semver==3.0.4
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.37.2
sympy==1.14.0
tabulate==0.9.0
tbb==2021.13.1
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.2.1
torch==2.3.1
tqdm==4.67.1
transformers==4.41.2
trio==0.30.0
trio-websocket==0.12.2
tritonclient==2.60.0
typer==0.9.4
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.29.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
youtube-transcript-api==0.6.3
yt-dlp==2023.12.30
zipp==3.23.0
